updates for prev comments
