+++ client/package.json
@@ -27,7 +27,9 @@
     ],
     "rules": {
       "immutable/no-let": 2,
-      "immutable/no-mutation": 2
+      "immutable/no-mutation": 2,
+      "immutable/no-this" : 2,
+      "no-param-reassign" : 2
     }
   },
   "browserslist": [
