+++ client/src/App.js
@@ -2,7 +2,7 @@ import React, { Component } from 'react';
 import { ApolloProvider } from 'react-apollo';
 import ApolloClient from 'apollo-boost';
 import './App.css';
-import Property from './components/Property'
+import Properties from './components/Properties'
 const client = new ApolloClient({
     uri : 'http://localhost:4000/graphql'
 });
@@ -15,7 +15,7 @@ class App extends Component {
 
          <div className="App">
          
-            <Property></Property>
+            <Properties></Properties>
 
          </div>
